embeddings/requirements.txt
# CLIP Embedding Service Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
transformers>=4.35.0
torch>=2.1.0
Pillow>=10.0.0
python-multipart>=0.0.6

